backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic==2.10.0
pydantic-settings==2.6.0
resend==0.8.0
groq>=1.1.2
email-validator==2.3.0
slowapi==0.1.9
pytest==7.4.3
httpx==0.26.0
APScheduler==3.10.4
reportlab==4.0.9
